frontend-app-api: move graph/ to top level
