+++ packages/frontend-app-api/src/graph/instantiateAppNodeTree.ts
