+++ packages/frontend-app-api/src/graph/readAppExtensionsConfig.test.ts
