+++ packages/frontend-app-api/src/graph/readAppExtensionsConfig.ts
