+++ packages/frontend-app-api/src/graph/resolveAppGraph.ts
